refactor: move WalsOpener to wal crate
