+++ analytic_engine/src/tests/read_write_test.rs
@@ -18,9 +18,9 @@ use std::{thread, time};
 
 use common_types::time::Timestamp;
 use logger::info;
+use wal::manager::WalsOpener;
 
 use crate::{
-    setup::WalsOpener,
     table_options,
     tests::util::{self, memory_ctxs, rocksdb_ctxs, EngineBuildContext, TestContext, TestEnv},
 };
